🚚 Rename pages/error -> pages/not-found
