--- src/app/route/error.routes.js
@@ -1,8 +1,0 @@
-import { ErrorPage } from "../../pages/errors";
-import { router } from "../../shared/libs";
-
-export const errorRoutes = () => {
-  router.addRoute("*", () => {
-    new ErrorPage().render();
-  });
-};
