+++ src/app/route/not-found.routes.js
@@ -1,0 +1,8 @@
+import { NotFoundPage } from "../../pages/not-found";
+import { router } from "../../shared/libs";
+
+export const notFoundRoutes = () => {
+  router.addRoute("*", () => {
+    new NotFoundPage().render();
+  });
+};
