--- src/pages/errors/index.js
@@ -1,1 +1,0 @@
-export { ErrorPage } from "./ErrorPage";
