+++ src/pages/not-found/index.js
@@ -1,0 +1,1 @@
+export { NotFoundPage } from "./NotFoundPage";
