+++ src/widgets/feed/constants/mock.constants.js
