+++ src/widgets/feed/feed.js
