+++ src/widgets/feed/index.js
