+++ src/widgets/footer/Footer.js
