+++ src/widgets/footer/index.js
