+++ src/widgets/header/NavigationBar.js
